+++ examples/resources/bunnynet_account_subuser/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_account_subuser.test john@example.com