+++ internal/provider/provider.go
@@ -109,6 +109,7 @@ func (p *BunnynetProvider) Configure(ctx context.Context, req provider.Configure
 
 func (p *BunnynetProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
+		NewAccountSubuserResource,
 		NewComputeContainerAppResource,
 		NewComputeContainerImageregistryResource,
 		NewComputeScriptResource,